SDK-691: Fixed test and update phpstorm integration
