+++ src/Presentation/Console/Commands/UpdateSdkCommand.php
@@ -7,26 +7,14 @@
 
 namespace SprykerSdk\Sdk\Presentation\Console\Commands;
 
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Helper\ProcessHelper;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Process\Process;
 
-class UpdateSdkCommand extends Command
+class UpdateSdkCommand extends AbstractUpdateCommand
 {
-    /**
-     * @var string
-     */
-    public const OPTION_CHECK_ONLY = 'check-only';
-
-    /**
-     * @var string
-     */
-    public const OPTION_NO_CHECK = 'no-check';
-
     /**
      * @var string
      */
@@ -36,11 +24,6 @@ class UpdateSdkCommand extends Command
      * @var string
      */
     protected string $sdkBasePath;
-
-    /**
-     * @var string
-     */
-    protected static $defaultDescription = 'Update Spryker SDK to latest version.';
 
     /**
      * @var \Symfony\Component\Console\Helper\ProcessHelper
@@ -60,25 +43,31 @@ class UpdateSdkCommand extends Command
     }
 
     /**
-     * @return void
+     * @param \Symfony\Component\Console\Input\ArgvInput $input
+     * @param \Symfony\Component\Console\Output\OutputInterface $output
+     *
+     * @return int
      */
-    protected function configure()
+    public function execute(InputInterface $input, OutputInterface $output): int
     {
-        parent::configure();
-        $this->addOption(
-            static::OPTION_CHECK_ONLY,
-            'c',
-            InputOption::VALUE_OPTIONAL,
-            'Only checks if the current version is up-to-date',
-            false,
-        );
-        $this->addOption(
-            static::OPTION_NO_CHECK,
-            null,
-            InputOption::VALUE_OPTIONAL,
-            'Only checks if the current version is up-to-date',
-            false,
-        );
+        return (int)!($this->runInstallBundles($output) || $this->runUpdate($input, $output));
+    }
+
+    /**
+     * @param \Symfony\Component\Console\Output\OutputInterface $output
+     *
+     * @return int
+     */
+    protected function runInstallBundles(OutputInterface $output): int
+    {
+        $application = $this->getApplication();
+
+        if (!$application) {
+            return static::FAILURE;
+        }
+        $application->setAutoExit(false);
+
+        return $application->run(new ArrayInput([InstallSdkBundlesCommand::getDefaultName()]), $output);
     }
 
     /**
@@ -87,28 +76,20 @@ class UpdateSdkCommand extends Command
      *
      * @return int
      */
-    protected function execute(InputInterface $input, OutputInterface $output): int
+    protected function runUpdate(InputInterface $input, OutputInterface $output): int
     {
-        $application = $this->getApplication();
-
-        if (!$application) {
-            return static::FAILURE;
-        }
-        $application->setAutoExit(false);
-        $application->run(new ArrayInput([InstallSdkBundlesCommand::getDefaultName()]), $output);
-
         $process = Process::fromShellCommandline(
             $this->sdkBasePath .
             '/bin/console ' .
-            InitCommand::NAME .
+            UpdateCommand::getDefaultName() .
             str_replace('\'' . static::$defaultName . '\'', '', (string)$input),
         )->setTty(true);
 
-        $this->processHelper->run(
+        $result = $this->processHelper->run(
             $output,
             [$process],
         );
 
-        return static::SUCCESS;
+        return (int)$result->getExitCode();
     }
 }
